// src/lib/types.ts
export type Generated<T> =
  T extends ColumnType<infer S, infer I, infer U>
    ? ColumnType<S, I | undefined, U>
    : ColumnType<T, T | undefined, T>;
// ...
export interface TblMonstername {
  MNDatum: Date | null;
  MNProjectID: string | null;
  MonsternameGUID: Generated<string>;
}

export interface TblMonsternamepunten {
  MPActief: Generated<boolean | null>;
  MPBreedte: number | null;
  MPConform: Generated<boolean | null>;
  MPGUID: Generated<string>;
  MPLengte: number | null;
  MPOmschrijving: string | null;
  MPVestigingID: string | null;
}

export interface TblMonsters {
  MonsterAnalyseID: string | null;
  MonsterCodeMedium: string | null;
  MonsterGUID: Generated<string>;
  MonsterMNID: string | null;
  MonsterMPID: string | null;
  MonsterProjectID: string | null;
  MonsterStart: Date | null;
  MonsterStatus: string | null;
  MonsterStop: Date | null;
}